pyproject.toml
[project]
name = "srttranlate"
version = "0.1.0"
description = ""
authors = [
    {name = "bowman",email = "fancycode@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai (>=1.82.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "srt (>=3.5.3,<4.0.0)",
    "tqdm (>=4.67.1,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
